Server/sumohaiku/requirements.txt
-e svn+https://code.djangoproject.com/svn/django/trunk/
sorl-thumbnail==11.05.2
django-pagination==1.0.7
PIL==1.1.7
fabric==1.2.0
django-uni-form==0.7.0
feedparser==5.1
south
